Update remove window action, move types from desktop slice to dedicated file, and add config.ts
