+++ src/features/Desktop/Desktop.tsx
@@ -1,9 +1,10 @@
-import React, {lazy, useCallback, Suspense, useMemo, useState} from 'react';
+import React, {lazy, useCallback, Suspense, useState} from 'react';
 import 'react-grid-layout/css/styles.css';
 import 'react-resizable/css/styles.css';
-import { Responsive, WidthProvider, Layouts } from 'react-grid-layout';
+import { Responsive, WidthProvider } from 'react-grid-layout';
 import { useAppSelector, useAppDispatch } from "../../app/hooks";
-import {LayoutBreakpoint, updateLayouts, DesktopUIWindow} from './DesktopSlice';
+import { updateLayouts } from './DesktopSlice';
+import { LayoutBreakpoint, DesktopUIWindow } from './types';
 import WindowComponent from '../../components/Window';
 import { componentLoader, ComponentNames } from '../../utils/componentLoader';
 import styles from './Desktop.module.css';
@@ -27,14 +28,16 @@ const Desktop: React.FC = () => {
 
     if (loader) {
       return lazy(() =>
-        loader().then((module) => ({ default: module.default as LazyComponentType }))
+        loader().then((module) => {
+          return { default: module.default as LazyComponentType };
+        })
       );
     } else {
       throw new Error(`Unknown component: ${componentName}`);
     }
   },[]);
 
-  const layouts: Layouts = useMemo(() => {
+  /*const layouts: Layouts = useMemo(() => {
     const layoutsData: Layouts = { lg: [], md: [], sm: [], };
     let xPosition = -1;
 
@@ -48,13 +51,13 @@ const Desktop: React.FC = () => {
       layoutsData.sm.push({ i: `${window.id}`, x: xPosition * 4, y: index, w: 1, h: 1 });
     });
     return layoutsData;
-  }, [windows]);
+  }, [windows]);*/
 
   return (
     <div className={styles.desktop}>
       <ResponsiveReactGridLayout
         cols={{ lg: 12, md: 8, sm: 2 }}
-        layouts={savedLayouts ?? layouts}
+        layouts={savedLayouts}
         breakpoints={{ lg: 1200, md: 996, sm: 768 }}
         onDragStop={(layout) => {
           dispatch(updateLayouts({ layout, breakpoint }));
