+++ src/features/Desktop/DesktopSlice.ts
@@ -1,36 +1,24 @@
 import {createSlice, PayloadAction} from '@reduxjs/toolkit';
 import { Layouts, Layout } from 'react-grid-layout';
-import { ComponentNames } from '../../utils/componentLoader';
-
-export type LayoutBreakpoint = 'lg' | 'md' | 'sm';
-
-export type NewWindowLayout = {
-  [key in LayoutBreakpoint]: Layout;
-};
-
-export interface DesktopUIWindow {
-  id: string;
-  name: string;
-  lazyLoadComponent?: ComponentNames;
-  layout?: NewWindowLayout;
-}
+import { DesktopUIWindow, LayoutBreakpoint } from './types';
+import { LOCAL_STORAGE_LAYOUT_KEY, LOCAL_STORAGE_DESKTOP_KEY } from './config';
 
 interface State {
   desktopWindows: DesktopUIWindow[];
-  layouts: Layouts | undefined;
+  layouts: Layouts;
 }
-
-const LOCAL_STORAGE_DESKTOP_KEY = 'desktopUI.desktop';
-const LOCAL_STORAGE_LAYOUT_KEY = 'desktopUI.layouts';
 
 const storedWindows = localStorage.getItem(LOCAL_STORAGE_DESKTOP_KEY);
 const desktopWindows = storedWindows ? JSON.parse(storedWindows) as DesktopUIWindow[] : [];
 
 const storedLayouts = localStorage.getItem(LOCAL_STORAGE_LAYOUT_KEY);
-const layouts = storedLayouts ? JSON.parse(storedLayouts) as Layouts : undefined;
+const layouts = storedLayouts ? JSON.parse(storedLayouts) as Layouts : { lg: [], md: [], sm: [] };
 
 const saveWindowsToLocalStorage = (desktopWindows: DesktopUIWindow[]) => {
   localStorage.setItem(LOCAL_STORAGE_DESKTOP_KEY, JSON.stringify(desktopWindows));
+}
+const saveLayoutsToLocalStorage = (layouts: Layouts) => {
+  localStorage.setItem(LOCAL_STORAGE_LAYOUT_KEY, JSON.stringify(layouts));
 }
 
 const initialState: State = {
@@ -47,6 +35,12 @@ export const DesktopSlice = createSlice({
       if (index !== -1) {
         state.desktopWindows.splice(index, 1);
         saveWindowsToLocalStorage(state.desktopWindows);
+
+        const breakpoints: LayoutBreakpoint[] = ['lg', 'md', 'sm'];
+        breakpoints.forEach((breakpoint) => {
+          state.layouts[breakpoint] = state.layouts[breakpoint].filter(layout => layout.i !== action.payload);
+        });
+        saveLayoutsToLocalStorage(state.layouts);
       }
     },
     addWindow: (state, action: PayloadAction<DesktopUIWindow>) => {
@@ -58,12 +52,10 @@ export const DesktopSlice = createSlice({
         const md: Layout = action.payload.layout?.md ?? { i: `${action.payload.id}`, x: 0, y: 0, w: 4, h: 2 };
         const sm: Layout = action.payload.layout?.sm ?? { i: `${action.payload.id}`, x: 0, y: 0, w: 1, h: 1 };
 
-        if (state.layouts) {
-          state.layouts.lg && state.layouts.lg.push(lg);
-          state.layouts.md && state.layouts.md.push(md);
-          state.layouts.sm && state.layouts.sm.push(sm);
-          localStorage.setItem(LOCAL_STORAGE_LAYOUT_KEY, JSON.stringify(state.layouts));
-        }
+        state.layouts.lg.push(lg);
+        state.layouts.md.push(md);
+        state.layouts.sm.push(sm);
+        saveLayoutsToLocalStorage(state.layouts);
       }
     },
     updateLayouts: (
@@ -74,7 +66,7 @@ export const DesktopSlice = createSlice({
         ...state.layouts,
         [action.payload.breakpoint]: action.payload.layout,
       };
-      localStorage.setItem(LOCAL_STORAGE_LAYOUT_KEY, JSON.stringify(state.layouts));
+      saveLayoutsToLocalStorage(state.layouts);
     },
   }
 });
