+++ src/features/Desktop/config.ts
@@ -1,0 +1,2 @@
+export const LOCAL_STORAGE_DESKTOP_KEY = 'desktopUI.desktop';
+export const LOCAL_STORAGE_LAYOUT_KEY = 'desktopUI.layouts';
